// frontend/sdk/models/HintDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Unpaused Spotify API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface HintDto
 */
export interface HintDto {
    /**
     * The round this hint was revealed
     * @type {number}
     * @memberof HintDto
     */
    round: number;
    /**
     * Hint type identifier (genre, decade, etc.)
     * @type {string}
     * @memberof HintDto
     */
    type: HintDtoTypeEnum;
    /**
     * Human-readable label for the hint
     * @type {string}
     * @memberof HintDto
     */
    label: string;
    /**
     * The hint value
     * @type {string}
     * @memberof HintDto
     */
    value: string;
}


/**
 * @export
 */
export const HintDtoTypeEnum = {
    Genre: 'GENRE',
    Decade: 'DECADE',
    Album: 'ALBUM',
    Popularity: 'POPULARITY'
} as const;
export type HintDtoTypeEnum = typeof HintDtoTypeEnum[keyof typeof HintDtoTypeEnum];


/**
 * Check if a given object implements the HintDto interface.
 */
export function instanceOfHintDto(value: object): value is HintDto {
    if (!('round' in value) || value['round'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('label' in value) || value['label'] === undefined) return false;
    if (!('value' in value) || value['value'] === undefined) return false;
    return true;
}

export function HintDtoFromJSON(json: any): HintDto {
    return HintDtoFromJSONTyped(json, false);
}

export function HintDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): HintDto {
    if (json == null) {
        return json;
    }
    return {
        
        'round': json['round'],
        'type': json['type'],
        'label': json['label'],
        'value': json['value'],
    };
}

export function HintDtoToJSON(json: any): HintDto {
    return HintDtoToJSONTyped(json, false);
}

export function HintDtoToJSONTyped(value?: HintDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'round': value['round'],
        'type': value['type'],
        'label': value['label'],
        'value': value['value'],
    };
}